Created a bootstrap.js file as new entry point
